Fix error message for malformed env config

+++ node_cli/operations/volume.py
@@ -20,6 +20,7 @@
 import logging
 import os
 import shutil
+import tempfile
 
 from node_cli.utils.helper import run_cmd
 from node_cli.utils.git_utils import sync_repo
@@ -32,6 +33,10 @@ from node_cli.configs import (
 )
 
 logger = logging.getLogger(__name__)
+
+
+class FilesystemExistsError(Exception):
+    pass
 
 
 def update_docker_lvmpy_env(env):
@@ -81,8 +86,34 @@ def docker_lvmpy_install(env):
     logger.info('docker-lvmpy installed')
 
 
-def prepare_device(block_device):
-    run_cmd(['mkfs.btrfs', block_device])
-    mountpoint = os.path.join(SKALE_STATE_DIR, 'schains')
-    os.makedirs(mountpoint)
+def get_block_device_filesystem(block_device):
+    with tempfile.TemporaryDirectory(dir=SKALE_STATE_DIR) as tempdir:
+        try:
+            run_cmd(['mount', block_device, tempdir.path])
+            r = run_cmd(['stat', '-f', '-c', '%T'])
+            return r.stdout.decode('utf-8')
+        finally:
+            run_cmd(['umount', block_device])
+
+
+def format_as_btrfs(block_device):
+    filesystem = get_block_device_filesystem(block_device)
+    logger.info('Found filesystem on the %s: %s', block_device, filesystem)
+    if filesystem == 'btrfs':
+        logger.info('Formatting %s as btrfs', block_device)
+        run_cmd(['mkfs.btrfs', block_device])
+    else:
+        raise FilesystemExistsError(f'{block_device} contains {filesystem}')
+
+
+def mount_device(block_device, mountpoint):
+    os.makedirs(mountpoint, exists_ok=True)
+    logger.info('Mounting %s as btrfs', block_device)
     run_cmd(['mount', block_device, mountpoint])
+
+
+def prepare_block_device(block_device):
+    mountpoint = os.path.join(SKALE_STATE_DIR, 'schains')
+    if not os.path.ismount(mountpoint):
+        format_as_btrfs(block_device)
+        mount_device(block_device, mountpoint)
